requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.4.1 # https://github.com/andymccurdy/redis-py
celery==4.4.2  # pyup: < 5.0  # https://github.com/celery/celery

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-form
django-redis==4.11.0  # https://github.com/niwinz/django-redis
channels==2.4.0 # https://github.com/django/channels
channels-redis==2.4.2
django-filter==2.2.0 # https://django-filter.readthedocs.io/en/master/
Pillow==7.1.1  # https://pillow.readthedocs.io/en/stable/

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
# Django webpack-loader
django-webpack-loader==0.7.0
